Represents grid as a map instead of as a vector
I couldn't figure out a better way to bundle the grid's height n (which I name :size) with it's state (called :tiles) given my current inexperience with Clojure. I need this number for many calculations and it doesn't make sense to calculate it over and over since it remains constant for a given grid. As I keep learning, perhaps I'll find a more idiomatic way to encapsulate the size of a grid with its state.