Improves performance of manhattan-distance
Although `cost` is now only called a single time per `solve`,
`manhattan-distance` is called numerous times by `slide`. Using the
most difficult 8 puzzle `[8 6 7 2 5 4 3 0 1]` as an example,
`manhattan-distance` is called a total of 188,784 times. Before this
commit, it took nearly 2 seconds to make all those calls. Now it takes
less than 40ms.

This is a nice performance gain, but the real bottleneck is the
implementation of the `slide` function. Completely rethinking how
`slide` works in `search` should reduce solution time of any 8 puzzle
to under a second and (hopefully) make 15 puzzles as tractable as can
be for IDA* with Manhattan distance as the heuristic.