Simplifies fee calculation for each search step
Each state added to the priority queue has a Manhattan distance one
less or one more than the previous one. There is a flat fee of 1 for
each move. This means that the net change for a move is either 2
(moved farther from goal) or 0 (moved closer to goal). This commit
eliminates the unneeded passing of `cost` and `fee` in the state's
vector and simplifies the logic of calculating a state's fee. No real
change in performance, but it makes the code easier to follow.