Uses directions and :dir metadata in search
This commit reduces the number of slides per state by first determining
which directions are valid for the current state (i.e., those that
result in a different state and don't undo the previous slide) and only
then calling `slide` on those directions. This results in a modest, but
noticeable overall speed increase. However, `directions` is slower than
I'd like for a function that is called in every pass in `search`. I'll
see if I can refactor it to be more efficient, or rethink how I'm doing
things so as not to call relatively expensive functions so much.