Memoizes goal state
This was by no means a major bottleneck of the solver, but it
illustrates exactly why this program is so slow. Before this commit,
`goal` and `tile-count` were being recalculated on every call (that's
anywhere from thousands to billions of unnecessary calculations!).
There is no need to calculate either more than once. Those calculations
add up to 100s of ms even for an eight puzzle. Memoizing `goal` (which
ends up covering `tile-count`), reduces this to just two 50 µs
calculations.

It's likely I can use the same reasoning to fix the major bottleneck —
all the needless calculation of tile/blank positions. It seems the
reason I am having so much trouble with this is that I am still not
comfortable (or sure) how to handle state the Clojure way. Ideally, I'd
want tile/blank positions to be attributes of the grid to which they
belong (calculated once and then just accessed). I know how I'd
accomplish this the OO way, but I'm still struggling to see how one
does this in FP. I think using `defrecord` to model a grid state might
be the way to go, but we'll see as I continue to learn.