Memoizes slide
Again, this is not the crux of the bottleneck. Ideally, I'd like to not memoize so much as it uses far more memory than really necessary. I suspect I still have a glitch somewhere in my IDA* implementation. It's either getting stuck in a dead end/endless cycle and/or I'm not updating the costs correctly.